Moved _diff_and_commit, _push and _grab_version to baserelease.py.
Thefirst was duplicated and the second and third may be handy for
other code too.

_grab_version is the basic implementation, and is overridden in the
prereleaser.
